pyproject.toml
[tool.poetry]
name = "auth0-firebase-token-exchange"
version = "1.0.0"
description = ""
authors = ["30ème Ciel <admin@30emeciel.fr>"]

[tool.poetry.dependencies]
python = "^3.9"
core = { git = "https://github.com/30emeciel/py-core.git", branch="4"}
# core = { path = "../py-core", develop = true }
# core = { path = "../py-core", develop = false }
python-box = "^5.3.0"
requests = "^2.24.0"
flask = "^1.1.2"
# force version to so wheel (binary format) is also available on Windows with PY3.9
# grpcio = "1.34.0"

[tool.poetry.dev-dependencies]
python-dotenv = "^0.17.0"
pytest = "^6.2.3"
pytest-mockito = "^0.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
